+++ CHANGELOG.md
@@ -22,6 +22,8 @@ as necessary. Empty sections will not end in the release notes.
 
 ### Commits
 
+## [0.102.4] Release (2025-01-31)
+
 ### Note
 
 - This release has no code changes except NOTICE file(s) clarifications.
@@ -1041,7 +1043,8 @@ as necessary. Empty sections will not end in the release notes.
 - Tests: Make `ITCassandraBackendFactory` less flaky (#7186)
 - IntelliJ: Exclude some more directories from indexing (#7181)
 
-[Unreleased]: https://github.com/projectnessie/nessie/compare/nessie-0.102.3...HEAD
+[Unreleased]: https://github.com/projectnessie/nessie/compare/nessie-0.102.4...HEAD
+[0.102.4]: https://github.com/projectnessie/nessie/compare/nessie-0.102.3...nessie-0.102.4
 [0.102.3]: https://github.com/projectnessie/nessie/compare/nessie-0.102.2...nessie-0.102.3
 [0.102.2]: https://github.com/projectnessie/nessie/compare/nessie-0.102.1...nessie-0.102.2
 [0.102.1]: https://github.com/projectnessie/nessie/compare/nessie-0.102.0...nessie-0.102.1